decorators: Refactor to return meaningful data
The functions being decorated do not return any explicit values, so the
decorators got refactored to return data that is meaningful. For
example, the benchmark decorator now returns the time taken by the
function being benchmarked instead of printing the result directly,
which would cause it to either show up above the board at the end, or
not at all if the screen gets cleared one last time (which is desirable
to avoid double-printing the board at the end).

Signed-off-by: Raymond L. Rivera <ray.l.rivera@gmail.com>